requirements.txt
# Core API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# OpenCV and image processing
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.1.0

# HTTP client for health checks
requests==2.31.0
